Add type definitions
This commit adds type definitions for `plugin-error`.
These types are adapted from the ones of `gulp-util`.
The main changes are that the types are more based on the
documentation instead of the implementation. It means that the `error`
property used internally is not longer exposed as an option (it matches
the tests). It also tries to prevent the use of the `plugin` option if
it will be overriden. It also adds support for inference of additional
properties from wrapped errors when the second argument is a real error.

This commit also adds some tests for the types. It just runs the
compiler with the `noEmit` option and checks for errors.

Closes gulpjs/plugin-error#5